Fixed differences between modules. Fixed poms, spring xml paths
