+++ api/src/main/webapp/WEB-INF/web.xml
@@ -1,44 +1,47 @@
-<!DOCTYPE web-app PUBLIC
- "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
- "http://java.sun.com/dtd/web-app_2_3.dtd" >
+<web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee
+                             http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd" version="3.1">
 
-<web-app>
-  <display-name>Fashion tips</display-name>
+    <display-name>Fashion tips</display-name>
 
-
-      <context-param>
+    <!--Load Spring Configuration-->
+    <context-param>
         <param-name>contextConfigLocation</param-name>
         <param-value>
-          WEB-INF/fashion-tips-security.xml
+            classpath:spring/spring-persistence.xml
+            classpath:spring/spring-service.xml
         </param-value>
     </context-param>
-
-     <filter>
-       <filter-name>springSecurityFilterChain</filter-name>
-       <filter-class>org.springframework.web.filter.DelegatingFilterProxy</filter-class>
-     </filter>
-     <filter-mapping>
-       <filter-name>springSecurityFilterChain</filter-name>
-       <url-pattern>/*</url-pattern>
-     </filter-mapping>
-
     <listener>
-      <listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
+        <listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
     </listener>
 
+    <!--Load Spring MVC Configuration-->
     <servlet>
-      <servlet-name>fashion-tips</servlet-name>
-      <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
-      <load-on-startup>1</load-on-startup>
+        <servlet-name>mvc-dispatcher</servlet-name>
+        <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
+        <init-param>
+            <param-name>contextConfigLocation</param-name>
+            <param-value>classpath:spring/spring-mvc.xml</param-value>
+        </init-param>
+        <load-on-startup>1</load-on-startup>
     </servlet>
-
     <servlet-mapping>
-      <servlet-name>fashion-tips</servlet-name>
-      <url-pattern>/</url-pattern>
+        <servlet-name>mvc-dispatcher</servlet-name>
+        <url-pattern>/</url-pattern>
     </servlet-mapping>
 
-    <welcome-file-list>
-      <welcome-file>index.jsp</welcome-file>
-    </welcome-file-list>
+
+
+     <!--<filter>-->
+       <!--<filter-name>springSecurityFilterChain</filter-name>-->
+       <!--<filter-class>org.springframework.web.filter.DelegatingFilterProxy</filter-class>-->
+     <!--</filter>-->
+     <!--<filter-mapping>-->
+       <!--<filter-name>springSecurityFilterChain</filter-name>-->
+       <!--<url-pattern>/*</url-pattern>-->
+     <!--</filter-mapping>-->
+
 
 </web-app>
