--- persistence/main/java/com/bionic/edu/dao/UserDao.java
@@ -1,10 +1,0 @@
-package com.bionic.edu.dao;
-
-import com.bionic.edu.entity.User;
-
-/**
- * User Dao interface.
- */
-public interface UserDao extends GenericDao<User, Long> {
-
-}
