--- persistence/main/java/com/bionic/edu/entity/User.java
@@ -1,53 +1,0 @@
-package com.bionic.edu.entity;
-
-import javax.persistence.*;
-
-/**
- * Entity class for User.
- */
-@Entity
-@Table(name="users")
-public class User extends BaseEntity<Long> {
-
-
-    private String login;
-    private String email;
-    private String password;
-
-    public User() {
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "id=" + id +
-                ", login='" + login + '\'' +
-                ", email='" + email + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
-}
