+++ persistence/src/main/java/com/bionicuniversity/edu/fashiontips/dao/UserDao.java
@@ -1,0 +1,10 @@
+package com.bionicuniversity.edu.fashiontips.dao;
+
+import com.bionicuniversity.edu.fashiontips.entity.User;
+
+/**
+ * User Dao interface.
+ */
+public interface UserDao extends GenericDao<User, Long> {
+
+}
