+++ persistence/src/main/java/com/bionicuniversity/edu/fashiontips/dao/impl/UserDaoImpl.java
@@ -1,0 +1,12 @@
+package com.bionicuniversity.edu.fashiontips.dao.impl;
+
+import com.bionicuniversity.edu.fashiontips.dao.UserDao;
+import com.bionicuniversity.edu.fashiontips.entity.User;
+import org.springframework.stereotype.Repository;
+
+/**
+ * UserDao implementation.
+ */
+@Repository
+public class UserDaoImpl extends GenericDaoImpl<User, Long> implements UserDao{
+}
