--- service/main/java/com/bionic/edu/service/UserService.java
@@ -1,10 +1,0 @@
-package com.bionic.edu.service;
-
-import com.bionic.edu.entity.User;
-
-/**
- * Created by Sergiy on 18.11.2015
- * Project: fashion-tips
- */
-public interface UserService extends GenericService<User, Long>{
-}
