--- service/main/java/com/bionic/edu/service/impl/UserServiceImpl.java
@@ -1,11 +1,0 @@
-package com.bionic.edu.service.impl;
-
-import com.bionic.edu.entity.User;
-import com.bionic.edu.service.UserService;
-
-/**
- * Created by Sergiy on 18.11.2015
- * Project: fashion-tips
- */
-public class UserServiceImpl extends GenericServiceImpl<User, Long> implements UserService {
-}
