--- service/main/java/com/bionic/edu/service/util/exception/NotFoundException.java
@@ -1,27 +1,0 @@
-package com.bionic.edu.service.util.exception;
-
-/**
- * Created by Sergiy on 16.11.2015
- * Project: fashion-tips
- */
-
-public class NotFoundException extends RuntimeException{
-    public NotFoundException() {
-    }
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-
-    public NotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public NotFoundException(Throwable cause) {
-        super(cause);
-    }
-
-    public NotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
