+++ service/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>fashion-tips</artifactId>
-        <groupId>com.bionic.edu</groupId>
+        <groupId>com.bionicuniversity.edu</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
@@ -12,7 +12,7 @@
     <artifactId>fashion-tips-service</artifactId>
     <dependencies>
         <dependency>
-            <groupId>com.bionic.edu</groupId>
+            <groupId>com.bionicuniversity.edu</groupId>
             <artifactId>fashion-tips-persistence</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
