+++ service/src/main/java/com/bionicuniversity/edu/fashiontips/service/UserService.java
@@ -1,0 +1,10 @@
+package com.bionicuniversity.edu.fashiontips.service;
+
+import com.bionicuniversity.edu.fashiontips.entity.User;
+
+/**
+ * Created by Sergiy on 18.11.2015
+ * Project: fashion-tips
+ */
+public interface UserService extends GenericService<User, Long>{
+}
