+++ service/src/main/java/com/bionicuniversity/edu/fashiontips/service/impl/UserServiceImpl.java
@@ -1,0 +1,14 @@
+package com.bionicuniversity.edu.fashiontips.service.impl;
+
+import com.bionicuniversity.edu.fashiontips.entity.User;
+import com.bionicuniversity.edu.fashiontips.service.UserService;
+
+import javax.inject.Named;
+
+/**
+ * Created by Sergiy on 18.11.2015
+ * Project: fashion-tips
+ */
+@Named
+public class UserServiceImpl extends GenericServiceImpl<User, Long> implements UserService {
+}
