+++ service/src/main/java/com/bionicuniversity/edu/fashiontips/service/util/exception/NotFoundException.java
@@ -1,0 +1,27 @@
+package com.bionicuniversity.edu.fashiontips.service.util.exception;
+
+/**
+ * Created by Sergiy on 16.11.2015
+ * Project: fashion-tips
+ */
+
+public class NotFoundException extends RuntimeException{
+    public NotFoundException() {
+    }
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+
+    public NotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public NotFoundException(Throwable cause) {
+        super(cause);
+    }
+
+    public NotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
