+++ service/src/main/resources/spring/spring-service.xml
@@ -1,0 +1,13 @@
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:context="http://www.springframework.org/schema/context"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+       http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
+
+    <!--Enable bean configuration throw annotations-->
+    <context:annotation-config/>
+
+    <!--Create service beans-->
+    <context:component-scan base-package="com.bionicuniversity.edu.fashiontips.service"/>
+
+</beans>
